requirements.txt
pyTelegramBotAPI==4.14.0
aiohttp==3.9.5
beautifulsoup4==4.12.3
selenium==4.23.1
python-dotenv==1.0.1
pycryptodome==3.20.0
